refactor: move FTSDirective to a separate file
